Refactor Trading components for improved order management and UI consistency
- Replaced MakerTakerTabs with LimitOfferTab in TradingLayout to streamline the trading interface.
- Introduced MarketOfferTab to replace TakeOfferContent, enhancing clarity in order handling.
- Updated CreateOfferModal and TakeOfferModal to utilize MarketOfferTab, ensuring consistent order type handling.
- Enhanced type definitions to include filters for better order management across components.
