+++ pengui/src/features/trading/index.ts
@@ -16,7 +16,7 @@ export { default as OrderTooltip } from './ui/orderbook/OrderTooltip'
 export { default as TradingLayout } from './ui/layout/TradingLayout'
 export { default as TradingContent } from './ui/layout/TradingContent'
 export { default as FilterPanel } from './ui/layout/FilterPanel'
-export { default as MakerTakerTabs } from './ui/layout/MakerTakerTabs'
+export { default as LimitOfferTab } from './ui/layout/LimitOfferTab'
 export { default as CreateOfferModal } from './ui/modals/CreateOfferModal'
 export { default as CreateOfferForm } from './ui/make-offer/CreateOfferForm'
 export { default as TakeOfferModal } from './ui/modals/TakeOfferModal'
